changed enhancer spelling from enchancer, moved appropraite files from src/assets to src/lib folder and modified required file imports
