+++ src/components/previousUploadsGraph.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Bar as CBar } from 'react-chartjs-2';
 import PropTypes from 'prop-types';
-import colors from '../assets/colors';
+import colors from '../lib/colors';
 // Returns an object with unique keys for each present analysis and values
 //   corresponding to the count of each analysis for creating histograms
 export function countUploads(data) {
